SMS notifier
Using the Nexmo REST API we can send SMS messages to arbitrary phone
numbers.
Nexmo like to be paid for this service and are unlikely to provide it if
no balance remains, so log a warning when the balance gets low.

TODO: Log this remaining balance in the django DB so we can display a
warning on the web interface.