Show the remaining SMS balance on the alarms page
Retrieve the remaining balance on the Nexmo account and show it on the
alarm page. This seems like a good place, as it should be instantly
visible when enabling the alarm that the SMS notification won't work (if
the balance it too low).

Note that the request is currently synchronous. If it turns out it slows
the page down too much it should be possible to make an asynchronous
request for it (through jQuery).