Alarm: Distinguish the warning and alarm states.
The PLC has a warning state (blinking lights) on detection before the
'real' alarm. We want to wait to send an SMS until the 'real' alarm
activates. This gives the user time to deactivate it.

Also show the warning state on the web-ui although it's unlikely to
occur as web users must manually refresh.